pyproject.toml
[project]
name = "zendesk-skill"
version = "1.0.0"
description = "Zendesk CLI and MCP Server for Claude Code - Search tickets, manage support, and analyze metrics"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "mcp[cli]>=1.6.0",
    "prompt-security-utils",
]

[tool.uv.sources]
prompt-security-utils = { git = "https://github.com/andmarios/prompt-security-utils" }

[project.scripts]
zendesk = "zendesk_skill.cli:main_cli"
zendesk-mcp = "zendesk_skill.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zendesk_skill"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
